+++ src/main/java/kr/hhplus/be/server/interfaces/api/ApiControllerAdvice.java
@@ -1,0 +1,16 @@
+package kr.hhplus.be.server.interfaces.api;
+
+import kr.hhplus.be.server.interfaces.api.common.APIException;
+import kr.hhplus.be.server.interfaces.api.common.APIResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class ApiControllerAdvice {
+    @ExceptionHandler(APIException.class)
+    public ResponseEntity<APIResponse> handleAPIException(APIException apiException){
+        APIResponse apiResponse = apiException.getApiResponse();
+        return ResponseEntity.status(apiResponse.statusCode()).body(apiResponse);
+    }
+}
