+++ src/main/java/kr/hhplus/be/server/interfaces/api/payment/dto/PaySeatReservationRequestDTO.java
@@ -1,0 +1,7 @@
+package kr.hhplus.be.server.interfaces.api.payment.dto;
+
+public record PaySeatReservationRequestDTO(
+        long reservation_id
+) {
+
+}
