+++ src/main/java/kr/hhplus/be/server/interfaces/api/payment/dto/PaymentDTO.java
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.interfaces.api.payment.dto;
+
+public record PaymentDTO(
+        long user_id,
+        long amount
+) {
+    public static PaymentDTO getDefault(){ return new PaymentDTO(1L, 1_000_000L);};
+}
