+++ src/main/java/kr/hhplus/be/server/interfaces/api/token/dto/TokenDomainRequestBodyDTO.java
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.interfaces.api.token.dto;
+
+public class TokenDomainRequestBodyDTO {
+    public record TokenCreateRequestBodyDTO(
+        long date_id
+    ) {}
+
+}
