--- src/main/java/kr/hhplus/be/server/interfaces/concert/dto/ConcertDTO.java
@@ -1,20 +1,0 @@
-package kr.hhplus.be.server.interfaces.concert.dto;
-
-import java.util.Arrays;
-import java.util.List;
-
-
-public record ConcertDTO(
-        long concert_id,
-        String concert_name,
-        List<ConcertDateDTO> concertDates
-) {
-
-    public static ConcertDTO getDefault(){
-        List<ConcertDateDTO> defaultDates = Arrays.asList(
-                ConcertDateDTO.getDefault(),
-                ConcertDateDTO.getDefault()
-        );
-        return new ConcertDTO(1L, "콘서트 이름", defaultDates);
-    };
-}
