--- src/main/java/kr/hhplus/be/server/interfaces/concert/dto/ConcertSeatDTO.java
@@ -1,9 +1,0 @@
-package kr.hhplus.be.server.interfaces.concert.dto;
-
-public record ConcertSeatDTO(
-        long seatNum,
-        boolean isAvail,
-        long price
-) {
-    public static ConcertSeatDTO getDefault() { return new ConcertSeatDTO(1L, true, 160_000); }
-}
