--- src/main/java/kr/hhplus/be/server/interfaces/payment/dto/PaySeatReservationRequestDTO.java
@@ -1,7 +1,0 @@
-package kr.hhplus.be.server.interfaces.payment.dto;
-
-public record PaySeatReservationRequestDTO(
-        long reservation_id
-) {
-
-}
