--- src/main/java/kr/hhplus/be/server/interfaces/token/dto/TokenDTO.java
@@ -1,7 +1,0 @@
-package kr.hhplus.be.server.interfaces.token.dto;
-
-public record TokenDTO(
-        String token
-) {
-    public static TokenDTO getDefault() {return new TokenDTO("sample_jwt_token");};
-}
