--- src/main/java/kr/hhplus/be/server/interfaces/token/dto/TokenDomainRequestBodyDTO.java
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.interfaces.token.dto;
-
-public class TokenDomainRequestBodyDTO {
-    public record TokenCreateRequestBodyDTO(
-        long date_id
-    ) {}
-
-}
