+++ src/test/java/kr/hhplus/be/server/interfaces/api/common/APIResponseTest.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.interfaces.api.common;
+
+class APIResponseTest {
+
+}
