--- src/test/java/kr/hhplus/be/server/interfaces/common/APIResponseTest.java
@@ -1,7 +1,0 @@
-package kr.hhplus.be.server.interfaces.common;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class APIResponseTest {
-
-}
